Moved course controller files to core folder, deleted base template folder
