--- Source/Base-Template.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 18ec6c5b56f6e58478d2819925c62214
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
