--- Source/Base-Template/.commitlintrc.json
@@ -1,1 +1,0 @@
-{ "extends": ["@commitlint/config-conventional"] }
