--- Source/Base-Template/.dependencies.json
@@ -1,9 +1,0 @@
-{
-    "dependencies": [
-        {
-            "name": "Core",
-            "repository": "git@github.com:Innoactive/Creator.git",
-            "branch": "refs/tags/v1.3.1"
-        }
-    ]
-}
