--- Source/Base-Template/.gitattributes
@@ -1,129 +1,0 @@
-# Kindly borrowed (and extended)
-# from https://gist.github.com/nemotoo/b8a1c3a0f1225bb9231979f389fd4f3f
-
-# Set the default behavior, in case people don't have core.autocrlf set.
-* text=auto
-
-## Unity ##
-
-*.cs diff=csharp text
-*.cginc text
-*.shader text
-
-# linguist-generated leads to diffs not being shown on github
-# see https://github.com/github/linguist#generated-code and
-# https://stackoverflow.com/a/24382933/1142028
-*.mat binary merge=unityyamlmerge linguist-generated=true
-*.anim binary merge=unityyamlmerge linguist-generated=true
-*.unity binary merge=unityyamlmerge linguist-generated=true
-*.prefab binary merge=unityyamlmerge linguist-generated=true
-*.physicsMaterial2D binary merge=unityyamlmerge linguist-generated=true
-*.physicsMaterial binary merge=unityyamlmerge linguist-generated=true
-#*.asset binary merge=unityyamlmerge linguist-generated=true
-*.meta binary merge=unityyamlmerge linguist-generated=true
-*.controller binary merge=unityyamlmerge linguist-generated=true
-
-## git-lfs ##
-
-# Graphics
-*.jpg filter=lfs diff=lfs merge=lfs -text
-*.JPG filter=lfs diff=lfs merge=lfs -text
-*.jpeg filter=lfs diff=lfs merge=lfs -text
-*.JPEG filter=lfs diff=lfs merge=lfs -text
-*.png filter=lfs diff=lfs merge=lfs -text
-*.PNG filter=lfs diff=lfs merge=lfs -text
-*.psd filter=lfs diff=lfs merge=lfs -text
-*.PSD filter=lfs diff=lfs merge=lfs -text
-*.ai filter=lfs diff=lfs merge=lfs -text
-*.AI filter=lfs diff=lfs merge=lfs -text
-*.bmp filter=lfs diff=lfs merge=lfs -text
-*.BMP filter=lfs diff=lfs merge=lfs -text
-*.gif filter=lfs diff=lfs merge=lfs -text
-*.GIF filter=lfs diff=lfs merge=lfs -text
-*.img filter=lfs diff=lfs merge=lfs -text
-*.IMG filter=lfs diff=lfs merge=lfs -text
-*.tif filter=lfs diff=lfs merge=lfs -text
-*.TIF filter=lfs diff=lfs merge=lfs -text
-*.xcf filter=lfs diff=lfs merge=lfs -text
-*.XCV filter=lfs diff=lfs merge=lfs -text
-*.tga filter=lfs diff=lfs merge=lfs -text
-*.TGA filter=lfs diff=lfs merge=lfs -text
-
-# Videos
-*.mpg filter=lfs diff=lfs merge=lfs -text
-*.MPG filter=lfs diff=lfs merge=lfs -text
-*.mpeg filter=lfs diff=lfs merge=lfs -text
-*.MPEG filter=lfs diff=lfs merge=lfs -text
-*.mp4 filter=lfs diff=lfs merge=lfs -text
-*.MP4 filter=lfs diff=lfs merge=lfs -text
-*.avi filter=lfs diff=lfs merge=lfs -text
-*.AVI filter=lfs diff=lfs merge=lfs -text
-*.asf filter=lfs diff=lfs merge=lfs -text
-*.ASF filter=lfs diff=lfs merge=lfs -text
-*.mp4 filter=lfs diff=lfs merge=lfs -text
-*.MP4 filter=lfs diff=lfs merge=lfs -text
-*.mov filter=lfs diff=lfs merge=lfs -text
-*.MOV filter=lfs diff=lfs merge=lfs -text
-
-# Audio
-*.mp3 filter=lfs diff=lfs merge=lfs -text
-*.MP3 filter=lfs diff=lfs merge=lfs -text
-*.ogg filter=lfs diff=lfs merge=lfs -text
-*.OGG filter=lfs diff=lfs merge=lfs -text
-*.wav filter=lfs diff=lfs merge=lfs -text
-*.WAV filter=lfs diff=lfs merge=lfs -text
-
-# Models
-*.obj filter=lfs diff=lfs merge=lfs -text
-*.OBJ filter=lfs diff=lfs merge=lfs -text
-*.fbx filter=lfs diff=lfs merge=lfs -text
-*.FBX filter=lfs diff=lfs merge=lfs -text
-*.ply filter=lfs diff=lfs merge=lfs -text
-*.PLY filter=lfs diff=lfs merge=lfs -text
-*.raw filter=lfs diff=lfs merge=lfs -text
-*.RAW filter=lfs diff=lfs merge=lfs -text
-*.blend filter=lfs diff=lfs merge=lfs -text
-*.BLEND filter=lfs diff=lfs merge=lfs -text
-*.3ds filter=lfs diff=lfs merge=lfs -text
-*.3DS filter=lfs diff=lfs merge=lfs -text
-*.stl filter=lfs diff=lfs merge=lfs -text
-*.STL filter=lfs diff=lfs merge=lfs -text
-*.pmd filter=lfs diff=lfs merge=lfs -text
-*.PMD filter=lfs diff=lfs merge=lfs -text
-*.mesh filter=lfs diff=lfs merge=lfs -text
-*.MESH filter=lfs diff=lfs merge=lfs -text
-
-# Libs
-*.dll filter=lfs diff=lfs merge=lfs -text
-*.DLL filter=lfs diff=lfs merge=lfs -text
-*.a filter=lfs diff=lfs merge=lfs -text
-*.A filter=lfs diff=lfs merge=lfs -text
-*.so filter=lfs diff=lfs merge=lfs -text
-*.SO filter=lfs diff=lfs merge=lfs -text
-*.lib filter=lfs diff=lfs merge=lfs -text
-*.LIB filter=lfs diff=lfs merge=lfs -text
-*.jar filter=lfs diff=lfs merge=lfs -text
-*.JAR filter=lfs diff=lfs merge=lfs -text
-
-# Unity Light Build
-*.exr filter=lfs diff=lfs merge=lfs -text
-*.EXR filter=lfs diff=lfs merge=lfs -text
-LightingData.asset filter=lfs diff=lfs merge=lfs -text
-
-# Others
-*.pdf filter=lfs diff=lfs merge=lfs -text
-*.PDF filter=lfs diff=lfs merge=lfs -text
-*.r16 filter=lfs diff=lfs merge=lfs -text
-*.R16 filter=lfs diff=lfs merge=lfs -text
-*.zip filter=lfs diff=lfs merge=lfs -text
-*.unitypackage filter=lfs diff=lfs merge=lfs -text
-*.aif filter=lfs diff=lfs merge=lfs -text
-*.ttf filter=lfs diff=lfs merge=lfs -text
-*.TTF filter=lfs diff=lfs merge=lfs -text
-*.rns filter=lfs diff=lfs merge=lfs -text
-*.reason filter=lfs diff=lfs merge=lfs -text
-*.lxo filter=lfs diff=lfs merge=lfs -text
-
-# 3D Modelling tools
-*.sbs filter=lfs diff=lfs merge=lfs -text
-*.max filter=lfs diff=lfs merge=lfs -text
