--- Source/Base-Template/.github/ISSUE_TEMPLATE/BUG_REPORT.md
@@ -1,47 +1,0 @@
----
-name: Bug report
-about: Create a report to help us improve
-title: ''
-labels: ''
-assignees: ''
-
----
-
-### Describe the bug
-<!--- A clear and concise description of what the bug is. -->
-
-### Steps to Reproduce
-<!--- 
-Attempt to recreate the issue in a new project and provide steps to reproduce in a clear text step by step list. 
-Include code samples, errors and stacktraces if appropriate. 
-
-1. Go to '...'
-2. Click on '....'
-3. Scroll down to '....'
-4. See error 
--->
-
-1.
-1.
-1.
-1.
-1.
-
-
-### Expected behavior
-<!--- Briefly describe what you expect should happen. -->
-
-### Current behavior
-<!--- Briefly describe what is actually happening. -->
-
-### Screenshots
-<!--- If applicable, add screenshots to help explain your problem. -->
-
-### Context (Environment)
-<!--- 
-How has this issue affected you? What are you trying to accomplish? 
-Providing context helps us come up with a solution that is most useful in the real world 
--->
-
-### Possible Solution
-<!--- Not obligatory, but suggest a fix/reason for the bug, -->
