--- Source/Base-Template/.github/ISSUE_TEMPLATE/FEATURE_REQUEST.md
@@ -1,20 +1,0 @@
----
-name: Feature request
-about: Suggest an idea for this project
-title: ''
-labels: ''
-assignees: ''
-
----
-
-### Is the feature request related to a problem?
-<!--- A clear and concise description of what the problem is. Ex. I'm always frustrated when [...] -->
-
-### Proposal
-<!--- A clear and concise description of what you want to happen. -->
-
-### Alternatives
-<!--- A clear and concise description of any alternative solutions or features you've considered. -->
-
-### Additional context
-<!--- Add any other context or screenshots about the feature request here. -->
