--- Source/Base-Template/.github/PULL_REQUEST_TEMPLATE.md
@@ -1,38 +1,0 @@
-### Description
-<!---
-Please include a summary of the change and which issue is fixed. Please also include relevant motivation and context. List any dependencies that are required for this change. If this PR fixes an open issue, please link it.
--->
-
-**Fixes**: # (issue)
-
-### Type of change
-<!--- Please delete options that are not relevant. -->
-
-- [ ] Bug fix (non-breaking change which fixes an issue)
-- [ ] New feature (non-breaking change which adds functionality)
-- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
-- [ ] This change requires a documentation update
-
-### How Has This Been Tested?
-<!---
-Please describe the tests that you ran to verify your changes. Provide instructions so we can reproduce. Please also list any relevant details for your test configuration
--->
-
-
-**Test Configuration**:
-
-
-### Aditional Notes
-<!--- Please provide any additonal information -->
-
-### Checklist
-<!--- Make sure your PR meets the following criteria before opening it. -->
-
-- [ ] My code follows the [Coding Conventions](#coding-conventions)
-- [ ] I have performed a self-review of my code
-- [ ] I have commented my code, particularly in hard-to-understand areas
-- [ ] I have made corresponding changes to the documentation
-- [ ] My changes generate no new warnings
-- [ ] I have added tests that prove my fix is effective or that my feature works
-- [ ] New and existing unit tests pass locally with my changes
-- [ ] Any dependent changes have been merged and published in downstream modules
