--- Source/Base-Template/.releaserc.json
@@ -1,12 +1,0 @@
-{
-    "branch": "master",
-    "plugins": [
-      "@semantic-release/commit-analyzer",
-      "@semantic-release/release-notes-generator",
-      "@semantic-release/github",
-      ["semantic-release-ado", {
-        "varName": "Version",
-        "setOnlyOnRelease": false
-      }]
-    ]
-  }
