--- Source/Base-Template/LICENSE.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 904357d19cdebe442935f18eb8c27f57
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
