--- Source/Base-Template/NOTICE
@@ -1,15 +1,0 @@
-Innoactive Basic Template
-Copyright 2013-2019 Innoactive GmbH
-
-This product includes software developed at Innoactive GmbH (http://innoactive.de), licensed under the Apache License, Version 2.0.
-
-This product contains software https://github.com/Unity-Technologies/PostProcessing by Unity Technologies ApS under the Unity Companion License.
-
-This product contains software https://fonts.google.com/specimen/Archivo?query=Archivo designed mainly by Omnibus-Type, licensed under the Open Font License.
-
-This product contains software https://fonts.google.com/specimen/Archivo+Narrow?query=Archivo designed mainly by Omnibus-Type, licensed under the Open Font License.
-
-The following is courtesy of our legal counsel:
-
-It is your responsibility to ensure that your use and/or transfer does not
-violate applicable laws.
