--- Source/Base-Template/NOTICE.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 066b3e8ec6cca43479d1abc596691f5f
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
