--- Source/Base-Template/README.md.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: a4167fd49c2caec4899c1f86af8dc0d6
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
