--- Source/Base-Template/Runtime.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 17b84ee3fbff00549876b6278df37a96
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
