--- Source/Base-Template/Runtime/VRBuilder.Template.asmdef
@@ -1,17 +1,0 @@
-{
-    "name": "VRBuilder.Template",
-    "references": [
-        "GUID:c8561f9de838ac04d8feeda695bc572d",
-        "GUID:42a7f63f1de938c4aacfe96b96e26d88",
-        "GUID:6055be8ebefd69e48b49212b09b47b2f"
-    ],
-    "includePlatforms": [],
-    "excludePlatforms": [],
-    "allowUnsafeCode": false,
-    "overrideReferences": false,
-    "precompiledReferences": [],
-    "autoReferenced": true,
-    "defineConstraints": [],
-    "versionDefines": [],
-    "noEngineReferences": false
-}
