--- Source/Base-Template/Runtime/VRBuilder.Template.asmdef.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 0bb33e9c24101f241afa02ef68c5b06b
-AssemblyDefinitionImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
