+++ Source/Core/Runtime/CourseController/AttachToTraineeView.cs.meta
