+++ Source/Core/Runtime/CourseController/CourseControllerSetup.cs.meta
